requirements.txt
Flask
Flask-CORS
Pillow
numpy
tensorflow
onnxruntime
gunicorn


